Backend/unimentor/urls.py
"""
URL configuration for unimentor project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

from rest_framework.routers import DefaultRouter
from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView

# Include app routers
from users.urls import router as users_router
from mentors.urls import router as mentors_router
from bookings.urls import router as bookings_router
from payments.urls import router as payments_router
from reviews.urls import router as reviews_router

router = DefaultRouter()
router.registry.extend(users_router.registry)
router.registry.extend(mentors_router.registry)
router.registry.extend(bookings_router.registry)
router.registry.extend(payments_router.registry)
router.registry.extend(reviews_router.registry)


urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/', include(router.urls)),
    path('api/users/', include('users.urls')),  # Include users URLs for OAuth endpoints
    path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
    path('api/docs/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
    # App-specific endpoints that might not use router could be included here later
]
