Backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
requests==2.31.0
